+++ test/new/GeometryTests/runtests.jl
@@ -4,6 +4,8 @@ using Test
 
 @testset "GridTopologies" begin include("GridTopologiesTests.jl") end
 
+@testset "UnstructuredGridTopologies" begin include("UnstructuredGridTopologiesTests.jl") end
+
 @testset "Triangulations" begin include("TriangulationsTests.jl") end
 
 @testset "ConformingTriangulations" begin include("ConformingTriangulationsTests.jl") end
